Separated Dummy file and made first assertion
